functional update and delete personal post

+++ src/main/java/controller/client/post/DeletePostController.java
@@ -1,6 +1,11 @@
 package controller.client.post;
 
+import bean.AccountModel;
+import bean.PostModel;
+import constant.SystemConstant;
 import service.IPostService;
+import utils.SessionUtil;
+import utils.UploadFileUtil;
 
 import javax.inject.Inject;
 import javax.servlet.ServletException;
@@ -22,8 +27,19 @@ public class DeletePostController extends HttpServlet {
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         try {
             Long postId = Long.parseLong(req.getPathInfo().substring(1));
-            postService.deleteByPostId(postId);
-            req.setAttribute("message", resourceBundle.getString("delete_success"));
+            PostModel postModel = postService.findByPostId(postId);
+            if (postId != null) {
+                AccountModel accountModel = (AccountModel) SessionUtil.getInstance().getValue(req,
+                        SystemConstant.ACCOUNTMODEL);
+                if (postModel.getUserModel().getUserId() != accountModel.getUser().getUserId()) {
+                    throw new Exception("not permission");
+                }
+                postService.deleteByPostId(postId);
+                UploadFileUtil.deleteFile(postModel.getLinkImages());
+                req.setAttribute("message", resourceBundle.getString("delete_success"));
+            } else {
+                throw new Exception("not permission");
+            }
         } catch (Exception e) {
             req.setAttribute("message", resourceBundle.getString("delete_failed"));
             e.printStackTrace();
